pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "ambient-api (>=1.5.7,<2.0.0)",
    "bech32 (>=1.2.0,<2.0.0)",
    "geographiclib (>=2.0,<3.0)",
    "flask-cors (>=5.0.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"